+++ R/bigquery-partition.R
@@ -152,7 +152,6 @@ bqRefreshPartitionData <- function(table,
 #' @param missing.dates dates for which to run this function for
 #' @param priority Default to INTERACTIVE
 #' @param use.legacy.sql Defaults to env variable if specified
-#' @inheritParams bqCreateTable
 bqTransformPartition <- function(table,
                                  file,
                                  ...,